pyproject.toml
[project]
name = "observ"
version = "0.1.0"
description = "AI tracing SDK for Observ"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Observ AI", email = "support@observ.ai"}
]
keywords = ["observ", "ai", "llm", "tracing", "observability", "anthropic", "openai", "mistral"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "anthropic>=0.73.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "mistralai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/observ-ai/observ-python"
Repository = "https://github.com/observ-ai/observ-python"
Documentation = "https://github.com/observ-ai/observ-python#readme"
Issues = "https://github.com/observ-ai/observ-python/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
